+++ packages/web-server/src/input.ts
@@ -60,21 +60,6 @@ export interface QueryParamOptions {
 }
 
 /**
- * Apply to a parameter to indicate that it represents a session parameter (ie foo in /bar?foo=1)
- * @param name 
- * @deprecated Use WebEvent.current. Will be removed in 4.0.0
- */
-export function Request(name? : string) {
-	return InputAnnotation.decorator({
-		validTargets: [ 'parameter' ],
-		allowMultiple: false
-	})({ 
-		type: 'request', 
-		name 
-	});
-}
-
-/**
  * Apply to a parameter to indicate that it represents a path parameter (ie 'thing' in /hello/:thing)
  * @param name 
  */